Refactor pruning proof build functions
